requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
timm>=0.9.0

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Phase 3: Advanced evaluation and monitoring
psutil>=5.9.0
tqdm>=4.65.0
optuna>=3.2.0
GPUtil>=1.4.0

# Additional utilities
Pillow>=9.5.0

# Utilities
tqdm>=4.65.0
Pillow>=10.0.0
requests>=2.31.0

# Phase 2: Advanced hyperparameter tuning
optuna>=3.4.0
psutil>=5.9.0

# Optional: Experiment tracking (can be disabled for Vast.ai)
wandb>=0.15.0
accelerate>=0.20.0 